Split out client to send EventHub messages
